[dbCore] package hibernate moved to dbCore, HibernateUtils renamed to SessionFactoryUtils to avoid clashing with the same named class in Spring.
