+++ src/main/java/ru/otus/vsh/knb/dbCore/hibernate/dao/AccountDaoHibernate.java
@@ -1,16 +1,10 @@
-package ru.otus.vsh.knb.hibernate.dao;
+package ru.otus.vsh.knb.dbCore.hibernate.dao;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import ru.otus.vsh.knb.dbCore.dao.AccountDao;
-import ru.otus.vsh.knb.dbCore.dao.PersonDao;
 import ru.otus.vsh.knb.dbCore.model.Account;
-import ru.otus.vsh.knb.dbCore.model.Person;
-import ru.otus.vsh.knb.hibernate.sessionmanager.SessionManagerHibernate;
-
-import javax.annotation.Nonnull;
-import java.util.List;
-import java.util.Optional;
+import ru.otus.vsh.knb.dbCore.hibernate.sessionmanager.SessionManagerHibernate;
 
 @Slf4j
 @Component
