+++ src/main/java/ru/otus/vsh/knb/dbCore/hibernate/dao/BetDaoHibernate.java
@@ -1,4 +1,4 @@
-package ru.otus.vsh.knb.hibernate.dao;
+package ru.otus.vsh.knb.dbCore.hibernate.dao;
 
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 import ru.otus.vsh.knb.dbCore.dao.BetDao;
 import ru.otus.vsh.knb.dbCore.model.Bet;
 import ru.otus.vsh.knb.dbCore.model.PersonsInGames;
-import ru.otus.vsh.knb.hibernate.sessionmanager.SessionManagerHibernate;
+import ru.otus.vsh.knb.dbCore.hibernate.sessionmanager.SessionManagerHibernate;
 
 import javax.annotation.Nonnull;
 import java.util.Collections;
