+++ src/main/java/ru/otus/vsh/knb/dbCore/hibernate/dao/GameDaoHibernate.java
@@ -1,4 +1,4 @@
-package ru.otus.vsh.knb.hibernate.dao;
+package ru.otus.vsh.knb.dbCore.hibernate.dao;
 
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -6,8 +6,8 @@ import org.hibernate.Session;
 import org.springframework.stereotype.Component;
 import ru.otus.vsh.knb.dbCore.dao.GameDao;
 import ru.otus.vsh.knb.dbCore.model.*;
-import ru.otus.vsh.knb.hibernate.sessionmanager.DatabaseSessionHibernate;
-import ru.otus.vsh.knb.hibernate.sessionmanager.SessionManagerHibernate;
+import ru.otus.vsh.knb.dbCore.hibernate.sessionmanager.DatabaseSessionHibernate;
+import ru.otus.vsh.knb.dbCore.hibernate.sessionmanager.SessionManagerHibernate;
 
 import javax.annotation.Nonnull;
 import java.util.Collections;
