+++ src/main/java/ru/otus/vsh/knb/dbCore/hibernate/dao/PersonDaoHibernate.java
@@ -1,12 +1,11 @@
-package ru.otus.vsh.knb.hibernate.dao;
+package ru.otus.vsh.knb.dbCore.hibernate.dao;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import ru.otus.vsh.knb.dbCore.dao.PersonDao;
 import ru.otus.vsh.knb.dbCore.model.Person;
-import ru.otus.vsh.knb.hibernate.sessionmanager.SessionManagerHibernate;
+import ru.otus.vsh.knb.dbCore.hibernate.sessionmanager.SessionManagerHibernate;
 
-import javax.annotation.Nonnull;
 import java.util.List;
 import java.util.Optional;
