+++ src/main/java/ru/otus/vsh/knb/dbCore/hibernate/sessionmanager/SessionManagerHibernate.java
@@ -1,4 +1,4 @@
-package ru.otus.vsh.knb.hibernate.sessionmanager;
+package ru.otus.vsh.knb.dbCore.hibernate.sessionmanager;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
